e2e: share terminal helpers; refocus terminal after seeding
seedSandboxFile writes through the R console against a remote server, which moved focus to the console so the typed rstudio command ran as R code (server-linux CI failure).
